web/requirements.txt
Django>=2.2.0,<2.3.0
psycopg2>=2.8.2,<2.9.0
gunicorn>=19.9.0,<19.10.0
djangorestframework>=3.9.0,<3.10.0
dj-rest-auth>=1.0.4,<1.1.0
django-allauth>=0.39.1,<0.40.0
django-cleanup>=4.0.0,<4.1.0
django-extensions>=2.2.9,<2.3.0
django-rest-swagger>=2.2.0,<2.3.0
flake8>=3.8.3,<3.9.0
Pillow>=7.1.0,<7.2.0
Faker>=4.1.2,<4.2.0
django-dbbackup>=3.3.0,<3.4.0
dropbox>=11.0.0,<11.1.0
django-storages>=1.11,<1.12
